gene_data_explorer/gene_data_explorer/requirements.txt
flask
pandas
configparser
flask-sqlalchemy
clustergrammer2
pymysql
cryptography
scipy
sklearn
scikit-learn
matplotlib
fastcluster
statsmodels==0.11.1
flask-login
oauthlib
requests 
pyopenssl
mysql
Flask-Bootstrap
Flask-WTF